Mantis #0000312
There was a fairly major problem with the rollback queries.

The problem was caused when an element in the where condition was changed by the update clause.
What was happening was that the rollback entries weren't being inserted because when the update was done and it altered the values that were being used to match against in the where clause the insert-select would not pickup the new entries.
eg this is an example update that this causes problems :
      UPDATE sq_asset_link SET sort_order = sort_order - 1
      WHERE majorid = '1' AND sort_order > '2'

because when the addRollbackEntry() did it's insert-select like this :
      INSERT INTO sq_rollback_asset_link ()
      SELECT ... FROM sq_asset_link WHERE majorid = '1' AND sort_order > '2'

you can see how the row which was sort_order = '3' won't be picked up by the insert-select.

As a side note, changed addRollbackEntry() to _addRollbackEntry() because it really is a private fn