A heap of changes to get this thing working, basically had to run it through bit by bit because it looked like there had been some changes to the data structures of some attributes and whoever did this didn't finish the job.
Removed the checking of username and email if use_email_validation is true. Instead in saveAttributes() if use_email_validation is true we now check to see if username and email are in the required_fields array and if not, put them in with a little message.

Added a new edit_interface_screen_locations.xml which removed the Optional Locations section so you can't set it in the backend, mainly because it was being processed anyway when the form was being submitted by the user with optional locations selected.

Also when a user clicked the link to validate the account creation and it was checking optional locations it was butchering the create_locations array it was using and when combined with another bug caused by the data structure being different to how it was being accessed in the code you would end up with your new users being created in the root folder.

Fixed the  _getCreateLocations() to have 2 arguments like the parent one in asset builder, also changed all references to it to pass new argument.

Changed the unused _getCreateLink() to the newly created _modifyCreateLink(), though as I said in asset builder commit this is still only going to be used on the creation of the first link between the new asset and whatever parent.