New Attribute Type parameter_map
Attributes of this type facilitate the storage and manipulation of
information on where on-the-fly values for certain asset parameters
should be drawn from.  An example of an asset parameter that you might
want an on-the-fly value for is an asset listing's root node.  The places
that you can tell the attribute to get the values from are GET, POST and
the id/name/short_name of the asset currently being painted on the frontend
(as distinct from assets being printed in nest-content areas etc, which may
be the ones whose parameters are being controlled).  The big idea here is
that when you want to make some aspect of an asset type adjustable
on-the-fly, you can just include an attribute of this type and let the user
specify where they want the values to come from.  The asset's code then just
asks this attribute for the runtime value.
~