New storage method for metadata. All fields now have individual rows for their values. Because of this - and the fact that you can only have 1 value for any given field with an asset (i.e. if it is 2+ schemas that are applied to an asset) - the metadata array has changed format.
The Old format was :
Array (
    [sections] => Array (
        [SECTION_NAME] => Array (
            [fields] => Array (
                [FIELD_NAME] => Array (
                    [value] => ...
                    [value_components] => Array()
                )
            )
        )
    )
)

The new format is :
Array (
    [FIELD ASSETID] = Array (
        [name]  => <field name>
        [value] => ....
    )
)

Another thing to note is that the value is actually the combined value and value_components
vars. Placed in a single string that is escaped properly. Use Metadata_Field::encodeValueString()
and Metadata_Field::decodeValueString() to take it to and from string to separated vars.

Another point to note is that the metadata_field_date value has changed from storing it's default
as an int to using ISO 8601.