Listing Engine commits.
The listing engine asset was created so that the Search Page (in particular) could benefit from the listing enhancements recently added to Asset Listing (as well as long-standing ones, like asset type formats), to provide a common interface between the two. This means that a lot of the Asset Listing contents are moved to the common Listing Engine.

Attribute sorting has also been added. Sorted attributes are able to be set for each set type through the new Asset Sorting screen when the 'sort by' is set to 'Asset Attributes...'. All sorting options are now available through this screen.

Still to be committed are the moving of keywords for "asset name linked" and "asset short name linked" from the Search Page to the Listing Engine.

Also, search page has inherited the listing engine's "displayed asset types" attribute, which allows the user to set whether the searchable types are inherited or not (previously searchable types always searched inherited types). This required small changes to the Search Manager's queries to handle this, which in turn required that similar changes be made to Search Folder. So, Search Folder also has this feature added.

A step-3/reconfigure is required for the upgrade of the search page and asset listing.