sq_ast_lookup_design table is no more. replaced with a more generic sq_ast_lookup_value This is too facilitate in the saving of certain information per-url for an asset (as opposed to per asset or per link) Examples of this will be the paint layouts when they come into play.
Upgrade procedure (also on forums) :
1. Run install/step_02.php to get new table
2. Run scripts/upgrade_design_lookup.php

Now be aware that upgrade_design_lookup.php attempts to fix a problem that was introduced because the designid field in sq_ast_lookup_design was part of the primary key when it shouldn't have been. This wasn't causing any problems for the live table but was for the lookup table.
The script just takes a best guess as to what design was being used at the time because it can't know for sure.
Basically it just takes the first one that is returned by the DB, because that is in essence what would happen if you put the system into rollback mode and viewed the site anyway.