This will fix the bug with "Login Key incorrect" error message that appeared sometimes on a login box even when the login/password were correct.
The bug was related to the Login Key (a generated unique value used as a hidden field
of the login forms to prevent people to log back in just by pressing the browser "Back"
button. It is also saved in the Session).

This Login Key was regenerated for each HTTP request. So if we had a two requests in a
row, the login key was regenerated and therefore the value stored in the hidden field of
the login box created by the first HTTP request became deprecated.

Practically you could observe that bug by opening the same web page twice and if you try
to log into the first page, you will get the "Login Key incorrect".

This new solution is not regenerating the login key for each request but only when the
user has logged in succesfully.