A new feature and lots of spring cleaning in the form.
New feature:  Form submissions now have a details screen where you can edit
the answers to questions, as long as those questions still exist in the form
that created the submission.  If the submitted value is no longer a valid
value for the question you can leave it intact, or specify a new, valid, value.
If new questions have been added to the form and they are required you can
still leave them blank when editing the submission.  You can access the
submission's details screen via the submission log screen of the form asset.

Tidying up has occurred mostly in form_question and its children.  There is
now an attribute is_required defined in form_question which is used or ignored
by child classes as appropriate.  Several functions in form_question have
been renamed for clarity, and remnants of the submission-as-an-array days
have been replaced with a single 'value' member variable in the form_question
class.  Upgrade functions are in the management classes of form_question (to
fill in the new 'is_required' atttribute from the old 'required' ones) and form
(to make sure content files are regenerated in light of the new function names).

Adherence to coding standards in these classes has been... improved.

I've tested this in all ways I can think of, but let me know if you notice
it's broken something.