Creation of a new metadata default value table, to (potentially) speed up handling of default values in metadata by removing asset calls. This table will have two fields - assetid and encoded value (ie, it is stored ALONG WITH any value components). Changes have been made to utilise this new default value table.
A step 2+3 will be required for the new default metadata value table. A script to populate it will be available today or tomorrow.

The other change is the ability to revert a metadata field to the default. If a value is explicitly set to the same value as the default, it will not change if the default changes. This provides a 'Use Default' checkbox which, when clicked, will ignore any user-entered value and remove the field's value from the metadata table - therefore reverting it to the default. Those fields set to the default will change when the default value of the metadata field in question is. This will apply to text, select and date metadata fields only. This will not apply to thesaurus fields as there is no concept of a default in that field.

This may not be fully localised just yet; if there are any stray strings then they will be corrected today.