added getKeywordReplacement function
It provides a new mechanism for keyword replacements by looking up the function named 'getKeywordNameKeywordReplacement' so an asset wishing to add support for 'best_ever' custom keyword
should implement a function getBestEverKeywordReplacement() which returns a replacement string

This function is currently being used only by the listing manager. Shortly most of the keyword replacement functions will be using it. This is a preferred way of adding keyword support to
an asset and should be used as following:

1. query the bodycopy for keywords used in it
2. loop through the keywords and call $this->getKeywordReplacement($keyword) on each (assuming $this is an asset that is supposed to provide the keyword replacement)
3. give the list of replacements to the bodycopy when it's being printed

Use of this function will make the front-end quicker by only looking up the keywords needed. Currently assets generate replacements for ALL possible keywords even if they're never used.