The WYSIWYG plugin class now has a function called 'print_plugin_pre_get_html()'.
This function can contain any javascript that is to be execute before the html is
retrieved from the editor. Useful for plugins that edit the html for editing purposes,
but don't wish for these changes to become part of the html.

It differs from print_plugin_get_html(), in that print_plugin_pre_get_html() allows you
to edit the actual editor, and the new html will be returned. In print_plugin_get_html()
the text has already been retrieved so any changed to the editor will be applied only
to the editor, and the returned text will be the old text.