Keyword system changes.
Previously getAssetKeywords() returned a big array with descriptions and replacements for every keyword even though they were not going to be used. Instead, use getAvailableKeywords() in order to get the list of keywords and their descriptions.

Replacement is done by getKeywordReplacement() function. Any assets which introduce their own keywords are responsible for providing proper replacements functions in their class.

If the asset has more than one bodycopy and needs to distinguish different sets of keywords for different bodycopies, the logic should be implemented in onRequestKeywords() function. Also, see onRequestKeywords() in listing_engine.

Note : getAssetKeywords() still stays in asset.inc not to break other packages which are currently using it. However, it's deprecated.