Keyword system changes.
This file is using the new keyword system. Also, minor snoopy fixes have done.

Note: function generateKeywordReplacements() in metadata manager.

Previously, if the given keyword doesn't have replacements, it returned the array with the key with the value of empty string. Now, it doesn't provide the key in the result, i.e. if there is no replacement for the given key, the key doesn't exist in the result array.