Ported overnight changes to cache_manager.inc to DEV.
These changes involve the addition of a functions loadCacheNoRef() and saveCacheNoRef(),
that allow the loading and saving respectively of assets from the cache without needing
a reference to the asset itself. These new functions require only the asset id and the type
code in order to perform the same functionality.