Additions to listing_engine: asset selections & targets.
Asset Selections are columns of checkboxes, radio buttons, drop-down lists or textboxes that can be assigned to a list of assets. Additionally, each checkbox column has the option of additionally being able to have a 'Check All' checkbox be printed globally and/or by group.
Asset Targets are buttons that send the form to selected pages. Working in tandem with selections, targets can be used to send values to another page for processing, as in the case of the ecommerce cart.