Ecommerce Overhaul
The Ecommerce cart has been completely overhauled, and now uses asset selections to update and remove cart contents. All cart operations are now performed by an external library (ecommerce/lib/ecommerce_common.inc).
Items can be added to the cart from external listing engine pages by setting up an asset selection on the external page and a dynamic parameter on the cart, setting both to using the same field name, eg. an Asset
 Selection called 'Purchase' matches up to a Dynamic Parameter assigned to the GET variable 'Purchase'.


Additions to listing_engine: asset selections & targets.

Asset Selections are columns of checkboxes, radio buttons, drop-down lists or textboxes that can be assigned to a list of assets. Additionally, each checkbox column has the option of additionally being able to have a 'Check All' checkbox be printed globally and/or by group.
Asset Targets are buttons that send the form to selected pages. Working in tandem with selections, targets can be used to send values to another page for processing, as in the case of the ecommerce cart.


Note: Requires Step 3 and Compile Locale.