Moved the checking that of the SQ_MS_*_[WHITE|BLACK]_LIST from Internal_Message::send() to Messaging_Service::getLogTypes() so checks can be done from outside the internal message and because it was already storing temp data in the messaging servive _tmp private var.
Also found out that there was a bug in the whether to log or not was being calculated.
It would affect the black listed items.

Say you had this in your conf file (which is a default install):
define('SQ_MS_LOG_TO_DB_WHITE_LIST', '*');
define('SQ_MS_LOG_TO_DB_BLACK_LIST', 'asset.attributes.fulllog.scalar
asset.workflow.invitation
asset.workflow.announce
install.*');

Even though install.* was blacklisted it was being logged to the DB. Basically if you
had '*' in the white list anything with '.*' ending was being ignored in the black list.


The reason for needed a separate function was so that a management classes update function could
check what the settings were for the log types it uses and send notices if it thought something
was not right.

eg.

class Uni_Greenwich_Cron_Job_Print_Job_Reminder_Management extends Asset_Management
{
..
	function update()
	{
		$ret_val = parent::update();
		if (!$ret_val) return $ret_val;

		// We are going to check with the step_3 what the message service will send the proper emails for our custom
		// internal message type
		$ms = &$GLOBALS['SQ_SYSTEM']->getMessagingService();
		$log_types = $ms->getLogTypes('uni_greenwich.print_job.reminder');

		// include asset to get locales (if they have been generated yet)
		$GLOBALS['SQ_SYSTEM']->am->includeAsset($this->getAssetType());

		if (empty($log_types['send_mail'])) {
			// really want to have send_mail
			trigger_localised_error('UNI_GRE_0007', E_USER_NOTICE);
		}
		if (empty($log_types['log_to_db'])) {
			// really want to have logging to the DB
			trigger_localised_error('UNI_GRE_0008', E_USER_NOTICE);
		}
		if (!empty($log_types['log_to_file'])) {
			// don't really need it to log to system log
			trigger_localised_error('UNI_GRE_0009', E_USER_NOTICE);
		}

		return $ret_val;


	}//end update()
..