Final additions to structured_drop_down.
This monster function now has the ability to limit link type to list.
It Omits assets that lie underneath invalid assets.
It allows a flat listing(in either drop-down or check box mode).
It allows an empty entry.

These were brought in line to fit in with the asset builder, and it will use
this now for dynamic create locations.

So many parameters, which is annoying, but many are just able to be defaulted.
Any help with reducing these parameters or different approaches welcome.