New version of the recreate_link_tree script.
The old script was using AM->getLinks() calls. This was very, very
slow and actually had problems because the form_email asset is a
bridge and only returns TYPE_2 links, which meant all the form
submissions in the systems not having entries in the link tree
table.

This script does what the original one used to reading directly from
the sq_ast_lnk table. Making it much quicker. Also because of the
way this works it uses much less memory to run.

BIG CHANGE to the script is that it now does NOT run the insert queries
it is just sending it to STDOUT. This again is for speed, especially on
big systems.

If you are using Postgresql it will output in COPY format because pgsql
inserts much quicker with that.

Usage: php scripts/recreate_link_tree.php . > tmp.sql