Changed getUserPrefs to be inline with changes made to dev a while ago.
When there is no user id,(ehich is usually, if not always from the command line), instead
of returning all preferences, the system still attempts to find the appropriate value
for the asset type and pref value given. This fixes problems with the import_files script
(Bug #887) and will aid with any other scripts that are creating or manipulating
assets that use user preferences.