Modified ldap->search to suppress PHP notices thrown when ldap_search, ldap_read and ldap_list find no results. The behaviour of this function is exactly the same ie. it will still return 0 when there are no results, but it will do so a little more gracefully than before.
This change will mostly be noticed when ldap users are deleted,
causing internal messages and workflow logs that reference
this user to produce warnings(See bug #860). These warnings will
now not appear, and the calling function will be left to handle the
null results of ldap_bridge::getAsset() (which it already did) using
an \'Unknown User\' substitution or similar.

Could also help with the problems outlined in bug (#150) but cannot
verify.