Stopped the orphaning of Cron Jobs.
Deleting a cron job previously would delete its link
to the cron manager, but not to the asset it applied to.
This was causing cron jobs to remain in the system, and not
able to be deleted. Also due to the changes to getActiveJobs()
in dev in Future status Cron Jobs, all linked jobs were
displayed, including these deleted ones. So it appeared that
future status assets were never able to be deleted.

Now, when there are no links between a cron job and the cron manager,
it is placed in the trash, and future status jobs no longer appear
on the details screen when they have been deleted. So all cron jobs are
happy and have parents again...even if they are the trash :(