Split printFrontend() into 2 functions: printFrontend() and printFrontendAsset($asset)
printFrontendAsset does everything printFrontend Used to do, but instead of being
restricted to printing itself only, printFrontendAsset accepts an asset object
to print. This allows assets to override printFrontend, and then pass a new asset
(for example a child page, or a page it wishes to redirect to) to printFrontendAsset.

printFrontend now simply passes $this to printFrontendAsset, so it is very easy to
override, but if it isn't it performs exactly as previously.