This is the first commit of the new asset pairing of page_multiple_page and page_multiple_page_page. These two assets combine to create a system that allows content to be spread across multiple pages.
The Multiple Page is a controller that contains 1 or more Multiple Page Pages
below it. Each Multiple Page Page is basically a bodycopy with a web path,
though its web paths are controlled by the parent asset. Any content entered
into each of the child pages is able to contain keywords that reference
other pages that exist in its system of pages. So things like next page
and previous are available. These keywords are also available to paint layouts,
which is the best way to use Multiple Pages.