New feature added to the internal messaging system, which allows the asset taglines to be used in the log interfaces. In the lang_messages.xml file, you would define a message body as follows:
	<body><![CDATA[<SQ_MSG_TAGLINE>%assetid%</SQ_MSG_TAGLINE>]]> has done XYZ action</body>

When the internal message is logged to the filesystem or sent via email, it replaces the SQ_MSG_TAGLINE element with "AssetName (#AssetID)". When it's
written to the database, you end up with "<SQ_MSG_TAGLINE>123</SQ_MSG_TAGLINE> has done XYZ action", which is then replaced with
get_asset_tag_line(123) when the message is retrieved with Messaging_Service::getLogs().