1. Bug fix in decoding metadata string when the metadata_field_text has value components.
Regular expression which strips off the white spaces didn't work properly, so the key of
component value became different with the original key. It caused an error that
the component value was not updated on asset's metadata screen although database was updated.

2. paintValueInterface() function is added to metadata_field_text_edit_fns.inc

This function is being called from the trigger in order to get the input interface
without calling backend outputter