Yet another modification of the thesaurus system. Lack of unique Ids on terms proved very impractical in cases where they were used to link to real assets and when needed to be renamed. The thesaurus database table structure is modified once again. Thesaurus terms now reside in their own table, similar to matrix assets. Link table now links term IDs together. Renaming a term is now easier. Linking a term to a real asset using the new shadow asset support is more predictable (because each term has only 1 ID now, previously it had as many IDs as it had parents). This property is also important for the upcoming event broadcasts on thesaurus terms.
Because this is a dev version, the only upgrade path from the previous dev version of thesaurus is to drop the db tables and re-run steps 2, 3 and locale. Everybody will need to do it, even if you did not use a thesaurus. If you
cannot afford to lose the existing dev thesaurus, export it, then re-import after upgrade.

The upgrade script from 3.6->3.8, distributed previously should still work as it didn't access the DB directly. The following queries are ONLY FOR 3.7 branch, not for upgrades from 3.6.

Queries that need to be run BEFORE step 2:
drop table sq_thes_term;
drop table sq_thes_lnk_tree;
drop sequence sq_thes_term_seq;