Changes to the listing engine grouping. Firstly, if the listing direction is UP the tree, the requirement that groupable parents have to be at or below the root node no longer applies (it makes no sense when all listed assets are already above the root!).
Also, if a depth/height limit applies, parent grouping assets are only selected when they lie in the correct position in relation to these limits. (For example, if we are going DOWN the tree and the depth limits are set to show all assets from 3 levels below the root node only, the first parent has to be 2 levels below, the second parent has to be 1 level below, etc.

The depth/height limit attributes have been moved from Asset Listing Page to the Listing Engine, although the ability to specify depth/height limits is still limited to Asset Listing Page. Step 3 required to transfer the attributes over