Fixed problem where the site asset was using its own design to print the index page, rather than the index page's design. This was caused by recent changes to use printFrontendAsset. With the new changes to printFrontendAsset that allow the passing in of a design, this problem has been fixed.
Instead of calling printFrontendAsset directly, site now reverts back
to paintAsset, which obtains the correct design to print. paintAsset()
has been changed so that once a correct design has been found, it then
calls printFrontendAsset, passing in the correct design. This results
in the site's index page being printed with its correct design when the root
URL of a site is accessed, rather than the site's design as previous.