New Content Type: Structure Tree
This new content type allows the display and editing of a tree,
useful for displaying asset structures, hierarchies or any situation
where different levels of importance need to be highlighted.

Preferences allow sys admins to customise what options users are
allowed to change on the tree interface, so users can be restricted from
deleting, highlighting, adding rows etc.

Preferences also allow the definition of classes to display tree nodes with,
which allows the styling of branches and node text of the trees on the frontend.