New Content Type Image
Allows the standardised display of images in content.
Preferences allow the customising of the CSS classes
to use for images, the fields to show on the create/edit
interface, and the default values for each field.

Also allows the automatic generation of inline previews or
new/same window links for thumbnails of the images.