Fixed problem with structured dropdown, and more specifically the dynamic creation dropdown in asset builder. The level limit or max depth option specified when displaying these dropdowns was working, but only when the dropdown was printed as a tree. When the dropdown was printed without structure selected, this caused all assets to be displayed.
Now, tree drawing function is called regardless of whether structured is
selected or not. There is a new argument, however, that determines whether
or not to print the actual visual representation of the structure. Basically
this allows the use of all of the tree drawing functionality, such as levels,
but allows the printing of this data as a flat dropdown option box.

NB. As its a complicated commit, coding standard fixes on the rest of this file
were not included in this commit, but will follow.