Numerous Remote Content improvements:
* Encoding of target url. This was done mainly because getvars in target urls were conflicting with getvars of the remote page itself, i.e. it was hard to distinguish who's the owner of the vars. As a side effect of this,
security is improved: Target urls are unreadable and it is impossible to just type an address in as a get var (unless it is properly encoded)

* Renaming the query var name to 'sq_content_src' from 'sq_remote_page_url' to hide the fact that remote content is being used. Coupled with the encoded URL, and given that the asset instance is not called 'remote content' it is
not obvious that the content comes from a remote source. Removing a redundant query var 'sq_remote_page_action' because 'sq_content_src' has enough info.

* Improving the way GET and POST vars are passed to the remote target. Scenarios like name=this&name=that are now properly handled and propagated, even though PHP itself doesn't handle this situation well. (NOTE: POST vars are
not supported when encoding is 'multipart/form-data', target page will not receive ANY variables. This shouldn't be a problem as Remote Content never handled file uploads)

minor optimisations and coding standards