Bug Fix #1097 the keyword %event_end_date% displays 1st Jan 1970 instead of the actuall event end date.
Since the recurring event may not have the end date of the first instance, the default value checking
is needed. So if the end date (for the first instance) is not set, it returns an empty string.

Also, %event_start_date% replacement type was timestamp format. Now, it's changed to human readable format.