This commit brings in the new roles system, which is essentially a contextual permission system. It allows you to specify "X user can perform Y role on Z asset," as well as some other funky things. To configure roles, go to the roles screen of any asset. To see what roles a particular user (or group) is able to perform, go to the role assignments screen for that user.
The roles system required a new DB view, which means that from now on if you want to query the permissions table you should query 'sq_vw_ast_perm'
instead of 'sq_ast_perm'...unless you're doing updates and deletions (or selects to find out what to delete), in which case you should continue to use
the 'sq_ast_perm' table.

There are also a few more metadata keywords available to show which roles are applied to an asset (much the same as permissions). Both the permissions
and role keywords allow you to expose a list of either the usernames, emails or linked emails. For permissions, these keywords are:

	%asset_[read|write|admin]_permission% --> Joe Blogs,John Doe
	%asset_[read|write|admin]_permission_email% --> joeblogs@example.com,johndoe@example.net
	%asset_[read|write|admin]_permission_linked% --> <a href="mailto:joeblogs@example.com">joeblogs@example.com</a>...

For roles, the keywords are:

	%asset_role_ROLEASSETID%
	%asset_role_ROLEASSETID_email%
	%asset_role_ROLEASSETID_email_linked%


Upgrade Requires:

	* Step 2 (create the new sq_ast_role table and the two new views)
	* Step 3
	* Re-parse designs