Added the ability to import and export designs.
Exporting:

Using the 'Import/Export' screen, the user is able
to specify a filename to export. This will tar up the design's
parse file, along with any associated files (not customisations)
into a file, and link a new asset with its contents under the design.

Importing:

Using either the file upload box to import a file from the user's computer,
or by selecting an existing file from the asset map, the design will create
a new link or linked file asset underneath the current design.

Application:

In the Current Files section, a list of any currently imported/exported design
archives are listed. Selecting the radio button next to any of these files
will apply its contents to the current design. It will apply the parse file,
create assets for any associated files, and will parse and process the design, so
that it is immediately usable without any extra operations. It will remove any
associated files and customisations that the design has before adding new assets.

Next to each file, there is also a link, which allows the user to download any
of the archives, so exported files can be transferred to other systems.

NB: Quite a big change, so Snoopy changes will follow for readibility and sanity