1. Any associated image varieties will be updated whenever the graph is updated.
2. New options has been added to the global settings screen
	- Graph can change the angle so that it draws horizontally
	- User can turn of the legend area
	- User can specify the padding for 4-sides of the axis(left, right, top and bottom)

3. Plot specific options. The different plot types have the different functionality. Previously
it was not possible to choose an option for a particular plot. Now plot settings section has been
divided into to sub sections so that the user can set the generic and specific options for a
particular plot.

At the moment Bar type plot can set the spacing between each bars in pixel.