Altered backup script, with some minor but important changes:
1.	Now it must be run from the System Root directory. Previously, it had this rule, but didnt enforce it.
	This simply ensures that the warning is displayed, and the execution is halted if it is run from anywhere else.
2.	Previously, the execution was stopped, based on any output from the TAR process being determined as an error.
	With the use of --exclude arguments, output is likely as the files they try to exclude may not exist, which
	results in a stat warning, but the archive is still created.
3.	Following on from 2, if there really is a problem creating the archive, not just the exclude warnings, then
	the error will be caught when the script tries to gzip the file. It will exit, as the gzip call will not
	be able to find the resulting archive.

This has not changed any of the functionality of the actual archiving, just now allows it to pass through until
the .gz file is created. If anyone wishes to implement some better exclude rules to the tar call from what is in
there (ie. ones that will not cause any errors when they don't match) please do. As for now I have left the previous
ones as they were.