New WYSIWYG Plugin, Definition List
Definition lists consist of type parts: a term and a description. The term is given by the DT element and
is restricted to inline content. The description is given by the DD element that contains block-level content.

The WYSIWYG Definition List is based on the same functionality in HTML editor, NVU.
The user can start to type by either pressing DT or DD button. Only difference between WYSIWYG and NVU the effect
of Enter key stroke. While NVU's Enter key will open a next element to type, the WYSIWYG uses Control-D shortcut for beginning
the next element. For instance, the input sequence :

Pressing DT

Type a term, 'Scott'

Control-D

Type a description, 'He is a boy.'

Control-D

Type a term, 'Scott2'

Control-D

Type a description, 'He is still a boy.'

will generate the html like the below.

<dl>
	<dt>Scott</dt>
	<dd>He is a boy.</dd>
	<dt>Scott2</dt>
	<dd>He is still a boy.</dd>
</dl>

To finish the definition list, you can press Control-D twice or press the same button from the menu as the tag name which your current cursor locates inside.
For example, if you just finished entering the description, pressing DD button will finish the plugin.

This plugin can also convert the existing text into the definition list. Each line which could be a paragraph or a single line with break tag will be
one DT or DD element in the list. At the moment this functionality works only for firefox. IE will be updated soon.