New WYSIWYG Plugin, Visual Aid
A visual indicator in the WYSIWYG which indicates the presence of the certain set of the tags. At the moment, it handles the lang attribute in SPAN tag,
the anchor tag and the table header(TH) and table cell(TD).

The developers should be aware of the fact that the visual aid is an extra html codes within the particular tag. In other words, they must
remember the presence of those code when they try to get the HTML content from the element or access innerHTML. If the plugin wants to modify
the content while visual aid is on, it must strip off the visual aid code before doing anything else. Or not many bad things can happen.