Feature Request #943:
Easier access to paint layout conditional keywords.
Following on from the addition of the begin and end keywords,
there is now an entry for the else keyword present in the dropdown
for each condition.

In addition, there is a new entry in the keyword dropdown, named
'Full Structure Condition X' for wach condtion, that will output all
three keywords for the condition, allowing the user to simply fill in the gaps.

Eg. 'Full Structure Condition Test' will insert the following into the editor:

%begin_condition_test%
%else_condition_test%
%end_condition_test%