filterAssetTypes() function is overriden in search_page.inc
The search manager uses the type_code column in the search indexing table to filter on asset types for performance reasons, so we don't
need to do any filtering. It simply returns the original search result back.

This change hasn't been made to 3-6-0 and 3-8-0 branches although their search managers are filtering asset types. When their search
page calls filterAssetTypes in listing_engine, it always returns an empty array so that the search result does not appear.

Bug fix 1214.