My last commit at squiz... FR 356 Added an attribute so that when a tag has multiple occurrences in the content, so as to determine if the content from all of the occurrences has to be returned or fromt he first occurence of the start tag to the firstr occurence of the end tag. For example in a situation like: <div> Tetsing this <div>Now</div> HERE</div> <div>Test that</div>
if you give the start tag as <div> and end tag as </div> then if you set the multioccurence attribute to true would give you all fo that content. But if you set it to false it would only give you
<div> Tetsing this <div>Now</div>
Which means the content starting from the first occurence of the start tag to the first occurence of the end tag.

Also fixed a few spelling mistakes in regards to 'tag_replacements' attribute