Bug #1216 Problems with search indexing and image variety
getAssetInfo() was not implemented in image.inc although it has its cute shadow asset kids underneath it. Then, asset_manager redirects
the request to the bridge (this time Image) and it happily return an empty array. It generated undefined index error.

Yes yes getAssetInfo() is not listed in bridge.inc. Maybe this is really the time to move on PHP5