Feature Request 573
Show the status of a DIV/TABLE in the edit content page.

The status description has been added to the Content DIV properties' tooltip on the edit contents screen so that the user can find out the status of bodycopy div without
expanding the asset map or changing the screen.