getRole() function fixes.
When a role is globally assigned, assetid, roleid and userid(zero) is added to role table indicating the global assignment.

This userid, zero, caused an Undefined Index error when the user group is assigned to that role since it attempts to get the
children of the usergroup by calling getChildren() with userid, zero.

Now foreach statement skips if it meets zero userid

One more place...