Bug Fix #1250
If you try to add a URL to a site (or special folder) and that URL is already applied, it throws the error but then fires the update
lookups HIPO anyway.

Previously, although one of the new URL is already used, the loop keeps running to process the rest of URLs so that the valid ones will
be updated. Now if any of newly entered is already in use, hipo will not update lookups.

It's not harsh... just type the right ones ^^